+++ example/main.tf
@@ -1,0 +1,3 @@
+resource "util_register" "example" {
+  set = "a1b2c3"
+}