+++ example/providers.tf
@@ -1,0 +1,9 @@
+terraform {
+  required_providers {
+    util = {
+      source = "poseidon/util"
+      #source = "terraform.localhost/poseidon/util"
+      version = "0.2.0"
+    }
+  }
+}