+++ util/provider.go
@@ -4,9 +4,12 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
-// Provider returns a config transpiler Provider.
+// Provider returns a util Provider.
 func Provider() *schema.Provider {
 	return &schema.Provider{
+		ResourcesMap: map[string]*schema.Resource{
+			"util_register": resourceRegister(),
+		},
 		DataSourcesMap: map[string]*schema.Resource{
 			"util_replace": datasourceReplace(),
 		},