+++ CHANGES.md
@@ -4,6 +4,10 @@ Notable changes between releases.
 
 ## Latest
 
+## v0.2.0
+
+* Add `util_register` resource for storing values
+
 ## v0.1.0
 
-* Add `util_replace` function
+* Add `util_replace` data source function